src/function/requirements.txt
google-cloud-secret-manager==2.16.4
google-cloud-kms==2.19.2
functions-framework==3.*
google-auth==2.22.0
